Move implementation-specific logic to separate packages
For example, The Guardian's logic now resides in one package.
