+++ src/main/scala/com/gu/kindlegen/Section.scala
@@ -1,12 +1,6 @@
 package com.gu.kindlegen
 
-import com.gu.contentapi.client.model.v1.Tag
 import com.gu.io.Link
-import com.gu.io.Link.AbsoluteURL
 
 
 case class Section(id: String, title: String, link: Link)
-
-object Section {
-  def from(tag: Tag): Section = Section(tag.id, tag.webTitle, AbsoluteURL.from(tag.webUrl))
-}
