--- src/test/scala/com/gu/kindlegen/ArticleSuite.scala
@@ -1,17 +1,0 @@
-package com.gu.kindlegen
-
-import org.scalatest.FlatSpec
-
-import com.gu.kindlegen.TestContent._
-
-class ArticleSuite extends FlatSpec {
-
-  val capiDate = ExampleDate
-  val ta = TestContent("", "", 2, "", "", capiDate, "", "", "", None)
-
-  ".apply" should "apply method with content as parameter" in {
-    val content = ta.toContent
-    val a = Article(content, ExampleGuardianProviderSettings)
-    assert(a.newspaperPageNumber === 2)
-  }
-}
