+++ src/test/scala/com/gu/kindlegen/MainSectionsBookBinderSpec.scala
@@ -4,7 +4,7 @@ import org.scalatest.FunSpec
 import org.scalatest.Inspectors._
 import org.scalatest.Matchers._
 
-import com.gu.kindlegen.TestContent.{ExampleLink, ExampleOffsetDate}
+import com.gu.kindlegen.TestData._
 
 
 class MainSectionsBookBinderSpec extends FunSpec {
