+++ src/test/scala/com/gu/kindlegen/SectionsManifestSpec.scala
@@ -5,16 +5,11 @@ import java.time.Instant
 import org.scalatest.FlatSpec
 import org.scalatest.Matchers._
 
+import com.gu.io.TempFiles
 import com.gu.io.Link.RelativePath
-import com.gu.kindlegen.TestContent._
+import com.gu.kindlegen.TestData._
 
-class SectionsManifestSpec extends FlatSpec {
-
-  private val capiDate = ExampleDate
-  private val time = Instant.now()
-
-  private val content = TestContent("", "", 1, "", "", capiDate, "", "", "", None).toContent
-  private val article = Article(content, ExampleGuardianProviderSettings)
+class SectionsManifestSpec extends FlatSpec with TempFiles {
 
   private val international = section("International", "theguardian/mainsection/international")
   private val topStories = section("Top Stories", "theguardian/mainsection/topstories")
@@ -22,6 +17,7 @@ class SectionsManifestSpec extends FlatSpec {
   private val sections = Seq(international, topStories, nonXmlChars)
 
   private lazy val bookSections = MainSectionsBookBinder.default.group(articles)
+  private val article = Article(section = null, 1, "", "", ExampleLink, ExampleOffsetDate, "", "", Nil, None)
   private val articles = Seq(
       (international, 1),
       (international, 2),
@@ -30,8 +26,9 @@ class SectionsManifestSpec extends FlatSpec {
       (topStories   , 4),
       (topStories   , 4),
       (nonXmlChars  , 5)
-    ).map {
-      case (section, page) => article.copy(newspaperPageNumber = page, section = section)
+    ).map { case (section, page) =>
+        Article(section = section, newspaperPageNumber = page,
+          "", "", ExampleLink, ExampleOffsetDate, "", "", Nil, None)
     }
 
   private def section(title: String, id: String) = {
@@ -40,9 +37,10 @@ class SectionsManifestSpec extends FlatSpec {
   }
 
   "SectionManifest.apply" should "convert a sequence of bookSections to a section Manifest (Contents page)" in {
-    val manifest = SectionsManifest("", ExamplePath, bookSections, time)
+    val time = Instant.now()
+    val manifest = SectionsManifest("", ExampleLink, bookSections, time)
     manifest.title shouldBe ""
-    manifest.link shouldBe ExamplePath
+    manifest.link shouldBe ExampleLink
     manifest.publicationDate shouldBe ExampleOffsetDate.toLocalDate
     manifest.buildInstant shouldBe time
